Added explicit casting to address ambiguous references

+++ src/test/java/com/cogpunk/math/IntegerOperatorTest.java
@@ -20,26 +20,26 @@ public class IntegerOperatorTest {
 	@Test
 	public void testAdd() {
 		
-		assertEquals(3, operator.add(1, 2));
-		assertEquals(1, operator.add(-1, 2));
+		assertEquals((Integer) 3, operator.add(1, 2));
+		assertEquals((Integer) 1, operator.add(-1, 2));
 		
 	}
 	
 	@Test
 	public void testSubtract() {
 		
-		assertEquals(1, operator.subtract(2, 1));
-		assertEquals(3, operator.subtract(2, -1));
-		assertEquals(-3, operator.subtract(2, 5));
+		assertEquals((Integer) 1, operator.subtract(2, 1));
+		assertEquals((Integer) 3, operator.subtract(2, -1));
+		assertEquals((Integer) (-3), operator.subtract(2, 5));
 		
 	}
 	
 	@Test
 	public void testMultiply() {
 		
-		assertEquals(2, operator.multiply(2, 1));
-		assertEquals(8, operator.multiply(2, 4));
-		assertEquals(-8, operator.multiply(2, -4));
+		assertEquals((Integer) 2, operator.multiply(2, 1));
+		assertEquals((Integer) 8, operator.multiply(2, 4));
+		assertEquals((Integer) (-8), operator.multiply(2, -4));
 		
 		
 	}
@@ -47,15 +47,15 @@ public class IntegerOperatorTest {
 	@Test
 	public void testDivide() {
 		
-		assertEquals(2, operator.divide(2, 1));
-		assertEquals(2, operator.divide(10, 4));
-		assertEquals(-3, operator.divide(22, -7));
+		assertEquals((Integer) 2, operator.divide(2, 1));
+		assertEquals((Integer) 2, operator.divide(10, 4));
+		assertEquals((Integer) (-3), operator.divide(22, -7));
 		
 	}
 	
 	@Test
 	public void testCast() {
-		assertEquals(3, operator.cast(3.5));
+		assertEquals((Integer) 3, operator.cast(3.5));
 	}
 	
 	@Test
