+++ src/test/java/com/cogpunk/math/probability/EventProbabilityProfileMultiplicationAggregationStrategyTest.java
@@ -17,7 +17,7 @@ public class EventProbabilityProfileMultiplicationAggregationStrategyTest {
 
 		Integer[] args = new Integer[] {1,2,3,4,5,6};
 		
-		assertEquals(720, new EventProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(Arrays.asList(args)));
+		assertEquals((Integer) 720, new EventProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(Arrays.asList(args)));
 		
 		
 		//assertEquals((Integer) 0, new ProbabilityProfileMultiplicationAggregationStrategy<Integer>(new IntegerOperator()).aggregate(new ArrayList<Integer>()));
